As per Stephens suggestion removed enums from request folder and moved ones used under cma/enums. Added input/output in log output. comments in build configuration for excluded folders
