+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/Cma.java
@@ -1,6 +1,6 @@
 package uk.gov.justice.laa.crime.cfecrime.cma;
 
-import uk.gov.justice.laa.crime.cfecrime.cma.request.CmaRequest;
+import uk.gov.justice.laa.crime.cfecrime.cma.enums.CmaRequest;
 import uk.gov.justice.laa.crime.cfecrime.cma.response.ValueList;
 
 public interface Cma {
