+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/StubCma.java
@@ -2,7 +2,7 @@ package uk.gov.justice.laa.crime.cfecrime.cma;
 
 import uk.gov.justice.laa.crime.cfecrime.cma.enums.FullAssessmentResult;
 import uk.gov.justice.laa.crime.cfecrime.cma.enums.MagCourtOutcome;
-import uk.gov.justice.laa.crime.cfecrime.cma.request.CmaRequest;
+import uk.gov.justice.laa.crime.cfecrime.cma.enums.CmaRequest;
 import uk.gov.justice.laa.crime.cfecrime.cma.response.*;
 import uk.gov.justice.laa.crime.cfecrime.cma.enums.CaseType;
 import uk.gov.justice.laa.crime.cfecrime.cma.enums.MeansTestOutcome;
