+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/enums/AssessmentType.java
@@ -1,0 +1,4 @@
+package uk.gov.justice.laa.crime.cfecrime.cma.enums;
+
+public enum AssessmentType {
+}
