+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/enums/CmaAssessment.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.cfecrime.cma.request;
+package uk.gov.justice.laa.crime.cfecrime.cma.enums;
 
 import java.util.Date;
 import java.util.List;
@@ -7,7 +7,7 @@ public class CmaAssessment {
     private AssessmentType type;
     private Date date;
     private CaseType caseType;
-    private CourtOutcome magistrateCourtOutcome;
+    private MagCourtOutcome magistrateCourtOutcome;
     private boolean hasPartner;
     private boolean hasPartnerContraryInterest;
     private EmploymentStatus employmentStatus;
