+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/enums/CmaRequest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.cfecrime.cma.request;
+package uk.gov.justice.laa.crime.cfecrime.cma.enums;
 
 public class CmaRequest {
     private CmaAssessment assessment;
