+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/enums/EmploymentStatus.java
@@ -1,0 +1,4 @@
+package uk.gov.justice.laa.crime.cfecrime.cma.enums;
+
+public enum EmploymentStatus {
+}
