--- src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/request/CourtOutcome.java
@@ -1,4 +1,0 @@
-package uk.gov.justice.laa.crime.cfecrime.cma.request;
-
-public enum CourtOutcome {
-}
