+++ src/main/java/uk/gov/justice/laa/crime/cfecrime/cma/response/MapResponse.java
@@ -7,8 +7,6 @@ import java.util.Map;
 
 public class MapResponse {
     private MapResponse(){
-        //MapResponse class
-        //to achieve code coverage 100%
     }
     public static class TreeNode {
         private final Map<String, Object> data;
